requirements.txt
asn1crypto==0.24.0
autopep8==1.4.3
boto3==1.9.121
botocore==1.12.121
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.6.1
cssmin==0.2.0
docutils==0.14
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsmin==2.2.2
MarkupSafe==1.1.0
ndg-httpsclient==0.5.1
peewee==3.8.2
peewee-db-evolve==3.7.0
psycopg2-binary==2.7.7
pusher==2.1.3
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-http-client==3.1.0
requests==2.21.0
s3transfer==0.2.0
sendgrid==5.6.0
six==1.12.0
urllib3==1.24.1
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.2.1
